Actualizo

+++ UNQfy/UNQfyAPI.js
@@ -41,12 +41,24 @@ router.use((req, res, next) => {
 
 // POST artist
 router.post('/artists', (req, res) => {
+    if(jsonIsValidForArtists(req.body)){
+        postArtist(req, res);
+    }else{
+        res.status(400).json({ status: 400, errorCode: "BAD_REQUEST" });
+    }    
+});
+
+function jsonIsValidForArtists(json){
+    return json.hasOwnProperty('name') && json.hasOwnProperty('country');
+}
+
+function postArtist(req, res){
     if (unqfy.getArtistByName(req.body.name) !== undefined) {
         res.status(409).json({ status: 409, errorCode: 'RESOURCE_ALREADY_EXISTS' });
     } else {
         addArtist(req, res);
     }
-});
+}
 
 function addArtist(req, res) {
     unqfy.addArtist({ name: req.body.name, country: req.body.country });
@@ -68,17 +80,25 @@ router.get('/artists/:id', (req, res) => {
         res.status(200);
         res.json(retArtist(artist));
     } else {
-        res.status(404).json({ "errorcode": "RESOURCE_NOT_FOUND" });
+        res.status(404).json({ status: 404, errorCode: "RESOURCE_NOT_FOUND" });
     }
 });
 
 // DELETE artist by ID
 router.delete('/artists/:id', (req, res) => {
-    unqfy.deleteArtistById(parseInt(req.params.id));
-    saveUNQfy(unqfy, 'estado');
-    res.status(200);
-    res.end();
+    if(artistExist(parseInt(req.params.id))){
+        unqfy.deleteArtistById(parseInt(req.params.id));
+        saveUNQfy(unqfy, 'estado');
+        res.status(200);
+        res.end();
+    }else{
+        res.status(404).json({ status: 404, errorCode: "RESOURCE_NOT_FOUND" });
+    }    
 });
+
+function artistExist(id){
+    return unqfy.getArtistById(id) !== undefined;
+}
 
 // GET artist by name
 router.get('/artists', (req, res) => {
@@ -96,7 +116,6 @@ function getArtist(name, res) {
         res.status(200);
         res.send(retArtists(artists));
     } else {
-        //res.status(404).json({ "errorcode": "RESOURCE_NOT_FOUND" });
         res.status(200).json([]);
     }
 }
@@ -118,13 +137,25 @@ function retArtist(artist){
 
 // POST album
 router.post('/albums', (req, res) => {
+    if(jsonIsValidForAlbums(req.body)){
+        tryPostAlbum(req, res);
+    }else{
+        res.status(400).json({ status: 400, errorCode: "BAD_REQUEST" });
+    }    
+});
+
+function jsonIsValidForAlbums(json){
+    return json.hasOwnProperty('name') && json.hasOwnProperty('year') && json.hasOwnProperty('artistId');
+}
+
+function tryPostAlbum(req, res){
     let artist = unqfy.getArtistById(req.body.artistId);
     if (artist === undefined) {
-        res.status(404).json({ "errorcode": "RELATED_RESOURCE_NOT_FOUND" });
+        res.status(404).json({ status: 404, errorCode: "RELATED_RESOURCE_NOT_FOUND" });
     } else {
         postAlbum(artist, req, res);
     }
-});
+}
 
 function postAlbum(artist, req, res) {
     if (artist.getAlbumByName(req.body.name) !== undefined) {
@@ -159,17 +190,25 @@ router.get('/albums/:id', (req, res) => {
             "tracks": album.tracks
         });
     } else {
-        res.status(404).json({ "errorcode": "RESOURCE_NOT_FOUND" });
+        res.status(404).json({ status: 404, errorCode: "RESOURCE_NOT_FOUND" });
     }
 });
 
 // DELETE album by ID
 router.delete('/albums/:id', (req, res) => {
-    unqfy.deleteAlbumById(parseInt(req.params.id));
-    saveUNQfy(unqfy, 'estado');
-    res.status(200);
-    res.end();
+    if(albumExist(parseInt(req.params.id))){
+        unqfy.deleteAlbumById(parseInt(req.params.id));
+        saveUNQfy(unqfy, 'estado');
+        res.status(200);
+        res.end();
+    }else{
+        res.status(404).json({ status: 404, errorCode: "RESOURCE_NOT_FOUND" });
+    }   
 });
+
+function albumExist(id){
+    return unqfy.getAlbumById(id) !== undefined;
+}
 
 // GET album by name
 router.get('/albums', (req, res) => {
@@ -203,6 +242,24 @@ function retAlbum(album){
     };
 }
 
+router.use('/*', (req, res) => {
+    res.status(404).json({ status: 404, errorCode: "RESOURCE_NOT_FOUND" });
+});
+
+//------------------------------------------------------------------
+
+function errorHandler(err, req, res, next) {    
+    if (err instanceof InvalidInputError){
+        res.status(err.status);
+        res.json({status: err.status, errorCode: err.errorCode});
+    } else if (err.type === 'entity.parse.failed'){
+        res.status(err.status);
+        res.json({status: err.status, errorCode: 'BAD_REQUEST'});
+    } else {
+        next(err);
+    }
+}
+
 //------------------------------------------------------------------
 
 app.use('/api', router);
@@ -212,9 +269,26 @@ app.use((req, res) => {
     res.status(404);
     res.json({status:404, errorCode: '"RESOURCE_NOT_FOUND"'})
 });
+*/
 
 app.use(errorHandler);
-*/
 
 app.listen(port);
 console.log('http://localhost:' + port + '/api');
+
+//------------------------------------------------------------------
+
+class APIError extends Error {
+    constructor(name, statusCode, errorCode, message = null) {
+        super(message || name);
+        this.name = name;
+        this.status = statusCode;
+        this.errorCode = errorCode;
+    }
+}
+
+class InvalidInputError extends APIError {
+    constructor() {
+        super('InvalidInputError', 400, 'INVALID_INPUT_DATA');
+    }
+}
